Completely change how movement works on the backend.
Now, there are two core movement functions: move() and turn_right().
Everything else is based on that, so for example move_left is
implemented by first detecting which direction the rover is facing,
turning right until it is facing left, and then moving forward.

This is currently only implemented on the backend (i.e. in Rust).
The UI still needs to be updated to reflect these changes. Also
"What's that Buzzing Sound" is currently broken and will need to
be re-worked.