Be more strict about passing challenges.
Now, in order to pass the challenge, your code needs to consider
every possible initial state (e.g. "left" and "right" in levels
which have asteroid detectors). If any possible initial state
results in the challenge not being passed, that is the one
that is selected and showing in the UI.