Don't update browser history.
This has some downsides but is the easiest way to force users to
see the loading screen the first time they visit. It also prevents
issues that can occur when using the back and forward buttons. The
most common issue is that sound would not play, but there may also
be problems with state management if a user unexpectedly visits a
page they were supposed to have already completed.