Refactor how simulation handles asteroids
Previously, asteroids were just "fuzzy" obstacles. "Fuzzy" here
means that the obstacle may or may not exist depending on which
possible initial state is ultimately selected when the simulation
is run. Fuzziness is a concept that originally extended to nearly
every part of the simulation state, but has long since only been
used for asteroids.

This change removes the general concept of "fuzziness" from the
simulation. Instead, asteroids are treated specially and a new
AsteroidWarning is generated automatically when an asteroid may
or may not appear in a given location. This allows us to simplify
both the frontend and backend code and sets us up for improved
asteroid animations.