Finally (hopefully) fix background images on Steam on Windows.
I think I finally tracked down the root of the problem. It has to
do with URL-escaping for the background-image property in CSS.
Ultimately, when running on Steam on Windows, the URL for each image
gets expanded and includes "Program Files (x86)" in the beginning.
Notably, this includes two parentheses. These parentheses can cause
problems in CSS unless they are escaped in a *very* specific way.
There are a handful of different ways of specifying URLs that work
totally fine, *except* for this specific edge case where there are
parentheses in the URL. For example not including the quotation marks
is valid CSS syntax but breaks if there are parentheses.

It has to be like this:

import starryBgImage from "../images/starry_bg.webp";
// later...
bgImage={`url("${starryBgImage}")`}

Most examples, guides, and StackOverflow posts don't need to worry
about parentheses in URLs, so that's why this was so hard to track
down and address. What makes it even harder is that apparently if
there is a problem with the CSS syntax, the browser just sort of
ignores the rule. At no point does React, Vite, or any other tool
raise an error. So I had very little to go on.